requirements.txt
google-generativeai
feedparser
requests
python-dotenv